memory leak fixes for the units dir - not all the way through yet though lovely way to spend a Friday afternoon :-)

+++ src/sbml/units/test/TestRunner.c
@@ -77,7 +77,7 @@ void
 setTestDataDirectory (void)
 {
   char *srcdir = getenv("srcdir");
-  int  length  = (srcdir == NULL) ? 0 : strlen(srcdir);
+  size_t  length  = (srcdir == NULL) ? 0 : strlen(srcdir);
 
 
   /**
